refactor: 관리자만 할수있는일 api 수정
